requirements.txt
Pillow==10.0.0
tqdm==4.66.1
